requirements.txt
marker-pdf
pydantic
groq
ragas
langchain-groq
langchain-community
langgraph
fastembed
python-dotenv
langchain